+++ delta/plugins/storage/src/main/scala/ch/epfl/bluebrain/nexus/delta/plugins/storage/files/Files.scala
@@ -366,7 +366,7 @@ final class Files(
       state   <- id match {
                    case Latest(_)        => log.stateOr(project, iri, notFound)
                    case Revision(_, rev) =>
-                     log.stateOr(project, iri, rev.toInt, notFound, RevisionNotFound)
+                     log.stateOr(project, iri, rev, notFound, RevisionNotFound)
                    case Tag(_, tag)      =>
                      log.stateOr(project, iri, tag, notFound, TagNotFound(tag))
                  }
