+++ delta/sdk/src/main/scala/ch/epfl/bluebrain/nexus/delta/sdk/resolvers/ResolversImpl.scala
@@ -142,7 +142,7 @@ final class ResolversImpl private (
       state   <- id match {
                    case Latest(_)        => log.stateOr(projectRef, iri, notFound)
                    case Revision(_, rev) =>
-                     log.stateOr(projectRef, iri, rev.toInt, notFound, RevisionNotFound)
+                     log.stateOr(projectRef, iri, rev, notFound, RevisionNotFound)
                    case Tag(_, tag)      =>
                      log.stateOr(projectRef, iri, tag, notFound, TagNotFound(tag))
                  }
